healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I2ec9f9a420ca61814d43c316b418ce94de3691bc

Former-commit-id: 6e7195bdf442c19d1c3f330b4e9bce3ce786f118

Former-commit-id: c713670f12bcb05706954fcd52a32ae6724f1d18
Former-commit-id: 51b32c8ed6a044bd432200fc9c11d50aae7c4162
Former-commit-id: bf912b575f17c2d66601748f866eebad093d42ff
Former-commit-id: fb835e0875b017299ac34320adca4cf0a0068a6f